modularized CreateButton
