+++ client/components/navbar.js
@@ -1,9 +1,8 @@
 import React from 'react'
 import {doSignOut} from '../firebase/auth'
-import {db} from '../firebase'
 import {withRouter} from 'react-router'
 import {Link} from 'react-router-dom'
-import {generateName} from '../3d/utilities/randomNameGenerator'
+import {CreateButton} from './CreateButton'
 
 class Navbar extends React.Component {
   constructor() {
@@ -11,56 +10,7 @@ class Navbar extends React.Component {
     this.state = {
       style: 'block'
     }
-    this.handleCreateWorld = this.handleCreateWorld.bind(this)
-    this.redirect = this.redirect.bind(this)
     this.signout = this.signout.bind(this)
-  }
-
-  async handleCreateWorld() {
-    const currentUser = this.props.currentUser
-    const worldsRef = db.ref('/worlds')
-    const newWorld = worldsRef.push()
-    const worldId = newWorld.key
-    const worldName = generateName()
-    if (currentUser) {
-      const userRef = await db.ref(`/users/${currentUser.uid}`).once('value')
-      const userData = userRef.val()
-      if (userData.worlds) {
-        const userWorldsRef = db.ref(`/users/${currentUser.uid}/worlds`)
-        userWorldsRef.update({
-          [worldId]: worldName
-        })
-      } else {
-        db.ref(`/users/${currentUser.uid}`).update({
-          worlds: {
-            [worldId]: worldName
-          }
-        })
-      }
-    }
-    newWorld.set({
-      id: worldId,
-      author: currentUser ? currentUser.displayName : 'guest',
-      name: worldName,
-      private: !!currentUser,
-      authorizedPlayers: [currentUser.displayName]
-    })
-    this.props.history.push(`/worlds/${worldId}`)
-    document.getElementById('dropdown').style.display = 'none'
-  }
-
-  toggleDropdown(event) {
-    const dropdown = event.target.nextElementSibling
-    if (dropdown.style.display === 'none' || dropdown.style.display === '') {
-      dropdown.style.display = 'flex'
-    } else {
-      dropdown.style.display = 'none'
-    }
-  }
-
-  redirect(event) {
-    this.props.history.push('/account')
-    document.getElementById('dropdown').style.display = 'none'
   }
 
   signout(event) {
@@ -81,17 +31,7 @@ class Navbar extends React.Component {
           </span>
         )}
         <div id="menu">
-          <div className="link-item dropdown" onClick={this.toggleDropdown}>
-            Create
-          </div>
-          <div id="dropdown">
-            <div className="link-item" onClick={this.handleCreateWorld}>
-              New world
-            </div>
-            <div className="link-item" onClick={this.redirect}>
-              Your creations
-            </div>
-          </div>
+          <CreateButton currentUser={currentUser} />
           <Link to="/worldlist">
             <div className="link-item">Explore</div>
           </Link>
